REF: Rename datasets/datasets.py to datasets/_loaders.py
